ViewModel setup is done
