+++ app/src/main/java/com/kannan/runningtrack/core/presentation/settings/SettingsFragment.kt
@@ -1,17 +1,16 @@
-package com.kannan.runningtrack.core.presentation.fragments
+package com.kannan.runningtrack.core.presentation.settings
 
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.Fragment
 import com.kannan.runningtrack.R
-import com.kannan.runningtrack.databinding.FragmentSetupBinding
+import com.kannan.runningtrack.databinding.FragmentSettingsBinding
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class SetupFragment : Fragment(R.layout.fragment_setup) {
-
+class SettingsFragment : Fragment(R.layout.fragment_settings) {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        val binding = FragmentSetupBinding.bind(view)
+        val binding = FragmentSettingsBinding.bind(view)
     }
 }
