+++ app/src/main/java/com/kannan/runningtrack/core/presentation/setup/SetupFragment.kt
@@ -1,16 +1,17 @@
-package com.kannan.runningtrack.core.presentation.fragments
+package com.kannan.runningtrack.core.presentation.setup
 
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.Fragment
 import com.kannan.runningtrack.R
-import com.kannan.runningtrack.databinding.FragmentStatisticBinding
+import com.kannan.runningtrack.databinding.FragmentSetupBinding
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class StatisticFragment : Fragment(R.layout.fragment_statistic) {
+class SetupFragment : Fragment(R.layout.fragment_setup) {
+
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        val binding = FragmentStatisticBinding.bind(view)
+        val binding = FragmentSetupBinding.bind(view)
     }
 }
