+++ app/src/main/java/com/kannan/runningtrack/core/presentation/statistic/StatisticFragment.kt
@@ -1,16 +1,16 @@
-package com.kannan.runningtrack.core.presentation.fragments
+package com.kannan.runningtrack.core.presentation.statistic
 
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.Fragment
 import com.kannan.runningtrack.R
-import com.kannan.runningtrack.databinding.FragmentTrackingBinding
+import com.kannan.runningtrack.databinding.FragmentStatisticBinding
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class TrackingFragment : Fragment(R.layout.fragment_tracking) {
+class StatisticFragment : Fragment(R.layout.fragment_statistic) {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        val binding = FragmentTrackingBinding.bind(view)
+        val binding = FragmentStatisticBinding.bind(view)
     }
 }
