+++ app/src/main/java/com/kannan/runningtrack/core/presentation/tracking/TrackingFragment.kt
@@ -1,16 +1,16 @@
-package com.kannan.runningtrack.core.presentation.fragments
+package com.kannan.runningtrack.core.presentation.tracking
 
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.Fragment
 import com.kannan.runningtrack.R
-import com.kannan.runningtrack.databinding.FragmentRunBinding
+import com.kannan.runningtrack.databinding.FragmentTrackingBinding
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class RunFragment : Fragment(R.layout.fragment_run) {
+class TrackingFragment : Fragment(R.layout.fragment_tracking) {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        val binding = FragmentRunBinding.bind(view)
+        val binding = FragmentTrackingBinding.bind(view)
     }
 }
